reorg and refactor
